authentication and authorization using JWT

+++ controllers/api/v1/posts_api.js
@@ -31,12 +31,19 @@ module.exports.index = async function(req, res){
 module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
-            post.remove();
+            if(post.user == req.user.id){
+                post.remove();
+            
 
-            await Comment.deleteMany({post: req.params.id});
-            return res.json(200, {
-                message: "Post and associated comments deleted successfully!"
+                await Comment.deleteMany({post: req.params.id});
+                return res.json(200, {
+                    message: "Post and associated comments deleted successfully!"
             });
+        }else{
+            return res.json(401, {
+                message: "You are not authorized to delete this post!"
+            });
+        }
     }catch(err){
         return res.json(500, {
             message: "Internal Server Error"
